backend/requirements.txt
fastapi
beautifulsoup4
uvicorn
openai
click
python-dotenv
sentence-transformers
PyPDF2
chromadb
langchain
tqdm
pandas
requests
langchain-huggingface
langchain-ollama
langchain-mongodb
pymongo
pypdf
torch>=2.0.0